src/workflow-trigger/requirements.txt
google-cloud-workflows==1.2.1